@@ -2,8 +2,8 @@ package sq.rogue.rosettadrone;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AlertDialog;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.appcompat.app.AlertDialog;
 import android.view.View;
 
 import static sq.rogue.rosettadrone.util.TYPE_DRONE_ID;
