@@ -1,8 +1,8 @@
 package sq.rogue.rosettadrone;
 
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import dji.common.error.DJIError;
